// src/types/user.type.ts
export interface IUser {
  idUser: Number;
  txtFirstName: String;
  txtMiddleName: String;
  txtLastName: String;
  txtVzid: String;
  txtEmailID: String;
  txtPhone: String;
  dtCreatedOn: Date;
  txtCreatedBy: String;
  dtUpdatedOn: Date;
  txtUpdatedBy: String;
  YnActive: Boolean;
  YnApproval: Boolean;
}